# Trie hash optimizations 