--- a/crates/primitives/Cargo.toml
+++ b/crates/primitives/Cargo.toml
@@ -20,7 +20,8 @@ revm-primitives = { workspace = true, features = ["serde"] }
 alloy-primitives = { workspace = true, features = ["rand", "rlp"] }
 alloy-rlp = { workspace = true, features = ["arrayvec"] }
 ethers-core = { workspace = true, default-features = false, optional = true }
-nybbles = { version = "0.1", features = ["serde", "rlp"] }
+# TODO: don't merge
+nybbles = { git = "https://github.com/dvush/nybbles", branch = "pop-idx-mut", features = ["serde", "rlp"] }
 
 # crypto
 secp256k1 = { workspace = true, features = ["global-context", "recovery"] }