--- a/crates/primitives/Cargo.toml
+++ b/crates/primitives/Cargo.toml
@@ -50,6 +50,7 @@ sucds = "~0.6"
 tempfile.workspace = true
 thiserror.workspace = true
 zstd = { version = "0.12", features = ["experimental"] }
+ahash.workspace = true
 
 # `test-utils` feature
 hash-db = { version = "~0.15", optional = true }