--- a/crates/storage/provider/Cargo.toml
+++ b/crates/storage/provider/Cargo.toml
@@ -36,6 +36,7 @@ pin-project.workspace = true
 parking_lot.workspace = true
 dashmap = { version = "5.5", features = ["inline"] }
 strum.workspace = true
+ahash.workspace = true
 
 # test-utils
 alloy-rlp = { workspace = true, optional = true }