This PR improves a performance of the trie when calculating hash.

In my benchmark it reduces time from 100ms to 90ms and this 10ms are more noticeable in implementation that caches db reads for the trie.

The most influential commits are:
* [cursor subnode full_key cache](https://github.com/paradigmxyz/reth/commit/6ba5d99e39cfe1fd04050e090c9f2d7cee87330b)
This key clone alone is mostly responsible for 10ms drop
* [AHash in trie](https://github.com/paradigmxyz/reth/commit/397f236e3f8b6b24ce38cfa04c734cd02a2bc6d1)
Recently hashbrown was removed from the repo and it made all hash maps / sets much slower. Its mostly due to hashbrown using ahash for hashing while std is using siphash. This only changes hashmaps in trie related code but other parts of reth can benefit from it too. 
* [prealloc children](https://github.com/paradigmxyz/reth/commit/bc60f6865c09a67709f05d228e25a3e3ab4018fc) is not that noticeable but improves `push_branch_node` a bit. (without it about 20% of the method is spent doing this collect) (moved to https://github.com/alloy-rs/trie/pull/1)