@onbjerg 2023-12-19
cc @DaniPopes this should be in alloy-trie, correct?