@DaniPopes 2023-12-20
> cc @DaniPopes this should be in alloy-trie, correct?

Only the primitives change. If you don't mind you can open a PR to https://github.com/alloy-rs/trie with the second commit @dvush