@DaniPopes 2023-12-20
Released [nybbles 0.1.2](https://docs.rs/nybbles/0.1.2/nybbles/)