@rkrasiuk 2023-12-26
@dvush in any follow-up perf PRs it would be great to see side by side comparison of flamegraphs / benches before and after the change. otherwise, it's tough to verify the performance boost from the changes. even some obvious "improvements" might degrade the performance as we've experienced in the past